dautism/requirements.txt
amqp==2.5.0
astroid==2.2.5
autopep8==1.4.4
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-configurations==2.1
django-polymorphic==2.1.2
django-stdimage==5.0.0
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
kombu==4.6.3
lazy-object-proxy==1.4.1
mccabe==0.6.1
newrelic==4.20.1.121
Pillow==6.1.0
progressbar2==3.42.0
psycopg2==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-logstash==0.4.6
python-utils==2.3.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
wrapt==1.11.2
